introduce *Alloc methods
Stmt.oneAlloc, Db.oneAlloc and Iterator.nextAlloc do the same thing as
Stmt.one, Db.one, Iterator.next respectively but they can allocate
memory.

This is useful when reading TEXT or BLOB columns because if you can't
allocate memory the only way to read these types is with an array which
means you must have an idea of the maximum size of the column.