workaround compiler bug
This workaround allows one to write this:

    const Foobar = struct {
    	foo: usize,
	bar: []const u8,
    };
    const row = try stmt.one(Foobar);

Where the first field in the struct (here `foo`) is either an integer type,
a float type or a boolean.
Moving the first field of this type to second position resolves the bug.

This is because readInt, readFloat and readBool have an empty error set and cause this
bug to trigger.

Removing the error altogether also triggers the bug so we define an
explicit error set.

If we actually want to return an error from these functions we'll simply
remove the workaround.

See https://github.com/ziglang/zig/issues/5149