ci: stop testing both memory and filesystem
We don't do much when choosing in memory or not, sqlite does all the
work of managing the data, so there's not much point for us testing both
modes.

This will also have the benefit of reducing the time needed to run the
CI, especially for aarch64 which is super slow.