sqlite: expose c import, add 'shared_cache' to init flags
Expose the C import to sqlite.h. Making a separate call to @cImport
outside of the library will cause Zig to regenerate all definitions in
sqlite.h. The regenerated definitions (i.e. structs, enums) would not
be equivalent to the definitions imported in by this library. This
causes problems in the case one wants to manually wrap SQLite structs,
pointers, and enums with the helpers provided in this library.

Added 'shared_cache' to init flags in order to allow having the same
backing table and statement cache shared amongst all connections pointed
to the same database file.