sqlite: interpret database path in init flags as uri
To avoid the need for introducing multiple sqlite.Mode's for addressing
all the different possible ways one may initialize a SQLite database,
enable the flag SQLITE_OPEN_URI by default.

This allows for initialization options which are not addressed by
InitFlags as of yet, such as the option to initialize a shared in-memory
SQLite database instance that may be shared across connections in the
same address space, to be set via. URI query parameters.

e.g.

  sqlite.Db.init({
    .mode = .{ .File = "file:hello.db?mode=memory&cache=shared" },
    .open_flags = .{ .create = true, .write = true },
    .threading_mode = .MultiThread,
  });