Stop using anytype in the public API.
The need for using `options: anytype` in readXYZ functions is so that
they can be used both when called by `one`/`next` or `oneAlloc`/`nextAlloc`.

In the first case there won't be an allocator member in the tuple, in
the latter there will be.

But, since the public API takes an explicit allocator argument in
`oneAlloc`/`nextAlloc` there's no need to take a `anytype` options in
the public API.

This commit changes the public API to always use `QueryOptions`.

This commit also adds a bunch of explicit comptime checks to validate
the options type passed to the readXYZ functions. Especially important
is checking the presence of the `allocator` field if the function
requires an allocator.

Finally, cleanup some stuff and reorder arguments in `readPointer`.