put the TODO comment at the end of the line instead
When using an explicit error set one must add the "Workaround" error in
their error set; putting the explanation as to why it exists at the end
of the line makes it so it is displayed directly in the output of the
compiler.

Before a user had to go look at the source code to understand why we
have this workaround.