rework execMulti
We don't have to expose the tail pointer to get execMulti to work: the
tail pointer can stay local to execMulti instead of polluting the
QueryOptions.

This is a breaking change but as far as I can tell people only use
execMulti, never the tail pointer for anything, so I guess it's fine.